### 2018-11-17 #1085 @whedon
[PRE REVIEW]: HRDS: A Python package for hierarchical raster datasets
**Submitting author:** @jhill1 (<a href="http://orcid.org/0000-0003-1340-4373">Jon Hill</a>)
**Repository:** <a href="https://github.com/EnvModellingGroup/hdrs" target ="_blank">https://github.com/EnvModellingGroup/hdrs</a>
**Version:** 0.1
