### 2018-11-17 #1085 @kthyng
@kbevers — I asked you about a different potential review about a month ago and you didn't have time then, but I wonder if you have time for a review now?

@edoddridge are you willing and able to review this JOSS submission? You can read more about the process [here](https://joss.readthedocs.io/en/latest/review_criteria.html).

Normally we ask for reviews in about 2 weeks, but for anyone in the US at least there is a holiday next week, so 3 weeks seems more reasonable.
