### 2018-11-18 5f1ae12
Adding support for nans in the raster data
