+++ hrds/raster_buffer.py
@@ -105,8 +105,10 @@ class CreateBuffer(object):
 
         # fill with edge value
         dist = np.full((ncols, nrows), 0.0)
-        # then fill in the middle
+        # then fill in the middle, 
+        # except where there is no data 
         dist[1:-1, 1:-1] = 1
+        dist[ self.raster.get_array == self.raster.ds.GetNoDataValue() ] = 0
         # calc euclidian distance and convert to 0 -> 1 scale
         dist = distance_transform_edt(dist, sampling=[dx[0], dx[1]])
         dist = dist / self.distance
