### 2018-11-18 9081e15
the raster holds the nodata val, not the data source
