### 2018-11-18 02a1386
Only deal with NaNs if buffer matches raster. Skipping test automatically
