### 2018-11-18 799220f
Adding test for dealing with nodata
