### 2018-11-19 5e13c5e
Adding code for min/max limits. Needs testing still
