### 2018-11-19 adcb476
Testing new min/max code. Tests passing locally
