### 2018-11-19 9e0df1a
Fixing lint errors
