### 2018-11-19 #1085 @edoddridge
I'd be happy to review this. I'll be away for the holiday most of this week, but can definitely do it within 3 weeks.
