### 2018-11-22 e4ad25e
Adding release badge
