### 2018-11-26 #1085 @kthyng
Ok thanks @kbevers!  And @edoddridge, great, and I'd like to get one more reviewer.

@ChrisBarker-NOAA  Are you or someone in your group willing and able to review this JOSS submission? You can read more about the process [here](https://joss.readthedocs.io/en/latest/review_criteria.html). We'd like the review within 2 weeks.
