### 2018-11-28 #1085 @ChrisBarker-NOAA
Kristen,

Doesn't look like too heavy a lift -- I can do it.

-CHB

On Mon, Nov 26, 2018 at 3:43 PM Kristen Thyng <notifications@github.com>
wrote:

> Ok thanks @kbevers <https://github.com/kbevers>! And @edoddridge
> <https://github.com/edoddridge>, great, and I'd like to get one more
> reviewer.
>
> @ChrisBarker-NOAA <https://github.com/ChrisBarker-NOAA> Are you or
> someone in your group willing and able to review this JOSS submission? You
> can read more about the process here
> <https://joss.readthedocs.io/en/latest/review_criteria.html>. We'd like
> the review within 2 weeks.
>
