### 2018-12-03 EnvModellingGroup/hrds#8 @edoddridge
Installation instructions - `pygdal` version
Submitted as part of a [review for JOSS](https://github.com/openjournals/joss-reviews/issues/1112).

The instructions include reference to a specific version of `pygdal`

> Install using pip, the correct version. Note you may need to increase the minor version number, e.g. from 2.1.3 to 2.1.3.3

> `pip install pygdal==2.1.3`

I needed to modify this to `pip install pygdal==2.2.3.3`. It might be better to state that the version number needs to match the output of `gdal-config --version` (with the possible addition of a minor version number), rather than pinning it to a specific version that will go stale quickly.
