### 2018-12-03 EnvModellingGroup/hrds#9 @edoddridge
Installation instructions - supported operating systems
Submitted as part of a [review for JOSS](https://github.com/openjournals/joss-reviews/issues/1112).

It's implicit in the `sudo apt-get install` instruction that the software is intended to be installed and run on a Linux distribution. It may be worth adding a statement about support, or lack thereof, for other systems.
