### 2018-12-03 EnvModellingGroup/hrds#10 @edoddridge
Examples - no example datasets included
Submitted as part of a [review for JOSS](https://github.com/openjournals/joss-reviews/issues/1112).

The readme includes two examples, but neither the bathymetry files required to run them, nor instructions on where to obtain the files.
