### 2018-12-03 EnvModellingGroup/hrds#11 @edoddridge
JOSS paper - GDAL citation
Submitted as part of a [review for JOSS](https://github.com/openjournals/joss-reviews/issues/1112).

The citation for GDAL doesn't follow the [recommendation provided by GDAL](http://trac.osgeo.org/gdal/wiki/FAQGeneral#HowdoIciteGDAL).
