### 2018-12-11 cf02d3b
Fixes #9
