+++ docs/papers.bib
@@ -1,9 +1,10 @@
 
 @MANUAL{GDALOGR2018,
-  title       = "{{GDAL/OGR} Geospatial Data Abstraction software Library}",
-  author      = "{GDAL/OGR contributors}",
+  title       = "{{GDAL/OGR} Geospatial Data Abstraction Library, Version 2.1.3.3}",
+  author      = {{GDAL Development Team}},
   institution = "Open Source Geospatial Foundation",
-  year        =  2018
+  year        =  2018,
+  url         = {http://www.gdal.org}, 
 }
 
 @ARTICLE{Avdis2018,
