### 2018-12-12 EnvModellingGroup/hrds#12 @edoddridge
Example dataset links
Submitted as part of a [review for JOSS](https://github.com/openjournals/joss-reviews/issues/1112).

https://github.com/EnvModellingGroup/hdrs/commit/174ca766867079aab03f7134e22bd357612cf3f9 introduces some very helpful description, but also has a few issues. The hyperlinks are missing for GEBCO and EMod and the OceanWise hyperlink is missing the exclamation mark that will make it render properly.
