### 2018-12-12 EnvModellingGroup/hrds#13 @edoddridge
GEBCO licencing
Submitted as part of a [review for JOSS](https://github.com/openjournals/joss-reviews/issues/1112).

While signing up to download GEBCO data so that I could actually run hdrs, I was presented with the following licence, which suggests that some GEBCO data probably could be included with hdrs. I realise that there are other reasons not to include data - repo size etc. - but having a couple of tiny example datasets would help users get started.

> GEBCO's gridded bathymetric data sets are placed in the public domain and may be used free of charge.

>Data within GEBCO's gridded bathymetric data sets are subject to copyright and database rights restrictions. Use
of GEBCO's gridded bathymetric data sets indicates that you accept the terms and conditions of use and
disclaimer information given below.

>Under these terms you are free to:

>Copy, publish, distribute and transmit the information

>Adapt the information

>Exploit the information commercially for example, by combining it with other information, or by including it in your own product or application
>Terms and conditions of use
>You must acknowledge the source of the data. A suitable form of attribution is given in the documentation that accompanies the data set.

>Ensure that you do not use the information in a way that suggests any official status or that GEBCO endorses you or your use of the information.

>Ensure that you do not mislead others or misrepresent the information or its source.
Disclaimer
>GEBCO's gridded bathymetric data sets are made available 'as is'. While every effort has been made to ensure reliability within the limits of present knowledge, the accuracy and completeness of GEBCO's gridded bathymetric data sets cannot be guaranteed. No responsibility can be accepted by those involved in their compilation or publication for any consequential loss, injury or damage arising from their use or for determining the fitness of the data for any particular use.

>Users should be aware that GEBCO's gridded bathymetric data sets are largely deep ocean products (based on trackline data from many different sources of varying quality and coverage) and do not include detailed bathymetry for shallow shelf waters. Although GEBCO's gridded bathymetric data sets are presented at 30 arc-second and one arc-minute intervals of latitude and longitude, this does not imply that knowledge is available on seafloor depth at this resolution. Users are advised to consult the accompanying data set documentation before using the data sets.

> GEBCO'S global bathymetric data sets shall not be used for navigation or for any other purpose involving safety at sea.

>Users are asked to report any problems encountered with the data as this feedback may result in the further improvement of GEBCO's data sets.
