### 2018-12-13 EnvModellingGroup/hrds#13 @jhill1
Excellent and well done on taking notice on the licence conditions! I've found a redistributable high resolution data set, so I should be able to create a full example. I'll commit and close this when done and also fix issue #12
Thanks!
