### 2018-12-13 6546e34
Adding base files for new example
