### 2018-12-13 EnvModellingGroup/hrds#14 @edoddridge
Test doc string
Submitted as part of a [review for JOSS](https://github.com/openjournals/joss-reviews/issues/1112).

In the buffer tests, something is wrong with one of the doc strings. It says a 10x10 grid with dx=0.2 and limits of (0,0) and (4,4). Either the size, the limits, or the dx must be wrong.

https://github.com/EnvModellingGroup/hdrs/blob/7c7a095cee17b3af79daa0b3fed874fee5e4e9eb/tests/test_buffer.py#L69-L72
