### 2018-12-13 EnvModellingGroup/hrds#15 @edoddridge
software name
Submitted as part of a [review for JOSS](https://github.com/openjournals/joss-reviews/issues/1112).

Is it `hrds` or `hdrs`? The repo is `hdrs`, but the directory name and JOSS paper both suggest `hrds`.
