### 2018-12-13 EnvModellingGroup/hrds#16 @edoddridge
Typos in JOSS paper
Submitted as part of a [review for JOSS](https://github.com/openjournals/joss-reviews/issues/1112).

A few minor suggestions for the paper.

1. Towards the end of the second paragraph there is a comma splice. A semicolon or full stop would be more grammatically traditional.
> There are limitations in this approach in that rasters cannot be partially overlapped, all but the base raster must be entirely contained within another raster.

2. "resolution" should probably get an L. Less of a typo, more of a stylistic point; I'd swap the second "in" for a "when".
> This software solves a particular problem when using multiscale numerical models, in that in using high resolution meshes, high resoution data is required, but for spatially limited regions.

3. "hierarchy" is misspelt. There is also an extraneous "to be used".
> By blending a heirachy of data sources, HRDS overcomes this problem and enables multiscale numerical problems to use spatially appropriate data to be used with minimal effort.

4. Is it "multi-scale" or "multiscale"? Both versions are used.
