### 2018-12-13 #1112 @edoddridge
I'm happy to say that I've finished my review of `HRDS` and can recommend it for publication. 👍

It is a well-written piece of code that solves a clearly stated problem. 'HRDS' provides a clean interface to mesh together multiple files with different resolutions and create manageable input files for variable resolution geophysical modelling.

While I've opened a number of issues in the code repository, they are all quite minor, and none of them should prevent publication of the paper. With perhaps one exception - it would be a shame to publish the paper before the typos are fixed.

I had very little difficulty obtaining a working version of the code. The dependencies were clearly identified. I encountered only one problem, and the documentation has now been updated to guide users around that issue. The test suite worked straight out of the box.

My only disappointment regarding the testing is the lack of a code coverage tool. The `.travis.yml` file contains some commented out lines for using coveralls to assess code coverage. I would recommend (re)implementing a code coverage tool.

The python code is laid out in a clear and logical way. I found it quite easy to work out what each piece of code does.

My biggest concern with this submission is the lack of a real-life worked example. However, @jhill1 has obtained the necessary data and has [committed to implementing one soon](https://github.com/EnvModellingGroup/hdrs/issues/13). Once the data is included, the test suite could be altered to use those datasets instead of the proprietary ones that @jhill1 uses locally.

The overarching documentation is currently limited to the readme. However, given the nature of the software, this is perfectly acceptable. API documentation is contained in docstrings for the various methods and classes. Having extensive docstrings means that Sphinx based documentation could be added in the future without too much extra effort.
