### 2018-12-19 1b51b4e
Adding code to make non-square rasters work via no-data. Tests swaped to new data, but need fixing
