+++ tests/test_buffer.py
@@ -41,13 +41,13 @@ class TestBufferCreator(unittest.TestCase):
             (on a 10x10 grid, not 4x4 as above...only so much space 
             in comments!)
             
-            LLC is 0,0 and upper right is 4,4. (hence dx is 0.2)
+            LLC is 0,0 and upper right is 4,4. (hence dx is 0.25)
             The data are stored in cell centres and we ask for a few coords.
             We check that the value is 1 at the distance, 0 at the boundary.
             Finally, test that the file is written and readable
             """
         rbuff = CreateBuffer(test_file_name1, 1.5)
-        point1 = [0.2, 3.85] # should return <0.2 (resolution)
+        point1 = [0.2, 3.85] # should return <0.25 (resolution)
         point2 = [1.7, 2] # should be 1
         point3 = [2, 2] # should be 1
         point4 = [0.85, 2] #should be 0.5
@@ -69,7 +69,7 @@ class TestBufferCreator(unittest.TestCase):
             (on a 10x10 grid, not 4x4 as above...only so much space 
             in comments!)
             
-            LLC is 0,0 and upper right is 4,4. (hence dx is 0.2)
+            LLC is 0,0 and upper right is 4,4. (hence dx is 0.25)
             The data are stored in cell centres and we ask for a few coords.
             We check that the value is 1 at the distance, 0 at the boundary.
             Finally, test that the file is written and readable
@@ -98,7 +98,7 @@ class TestBufferCreator(unittest.TestCase):
             in comments!). The slightly-left-of-centre has been 
             replaced by NaN
             
-            LLC is 0,0 and upper right is 4,4. (hence dx is 0.2)
+            LLC is 0,0 and upper right is 4,4. (hence dx is 0.25)
             The data are stored in cell centres and we ask for a few coords.
             We check that the value is 1 at the distance, 0 at the boundary
             and, for this test, the value is 0 in the centre.
