+++ hrds/raster_buffer.py
@@ -87,6 +87,25 @@ class CreateBuffer(object):
         dataset.FlushCache()
         return
 
+
+    def extend_mask(self, array, iterations):
+        # function to extend a mask array.
+        # Taken from: http://www.siafoo.net/snippet/82
+        # Copyright 2007 Regents University of California
+        # Written by David Isaacson under BSD licence
+        yLen,xLen = array.shape
+        output = array.copy()
+        for i in range(iterations):
+            for y in range(yLen):
+                for x in range(xLen):
+                    if (y > 0        and array[y-1,x]) or \
+                       (y < yLen - 1 and array[y+1,x]) or \
+                       (x > 0        and array[y,x-1]) or \
+                       (x < xLen - 1 and array[y,x+1]): output[y,x] = True
+            array = output.copy()
+    
+        return output
+
     def make_buffer(self, output_file):
         # make a raster of the same extent, but with
         # square resolution which is dependant on distance buffer
@@ -94,14 +113,16 @@ class CreateBuffer(object):
         if self.over is None:
             transform = self.raster.ds.GetGeoTransform()
             dx = [transform[1], -transform[5]]
+            nrows = self.raster.ds.RasterXSize
+            ncols = self.raster.ds.RasterYSize
         else:
             dx = [self.distance / self.over, self.distance / self.over]
+            # note this changes our extent if "over" is set
+            nrows = int(ceil((urc[0] - llc[0]) / dx[0]))
+            ncols = int(ceil((urc[1] - llc[1]) / dx[1]))
 
         llc = self.extent[1]
         urc = self.extent[3]
-        # note this changes our extent if "over" is set
-        nrows = int(ceil((urc[0] - llc[0]) / dx[0]))
-        ncols = int(ceil((urc[1] - llc[1]) / dx[1]))
 
         # fill with edge value
         dist = np.full((ncols, nrows), 0.0)
@@ -114,11 +135,16 @@ class CreateBuffer(object):
             # TODO this will only work if dist and orig_raster
             # are the same size
             dist[orig_raster == nodata] = 0
+            # we now extend this mask
+            mask = np.full((ncols,nrows),False)
+            mask[dist == 0] = True
+            mask = self.extend_mask(mask, 2)
+            dist[mask] = 0
         # calc euclidian distance and convert to 0 -> 1 scale
         dist = distance_transform_edt(dist, sampling=[dx[0], dx[1]])
         dist = dist / self.distance
         dist[dist > 1] = 1.0
 
         # create a suitable output filename
-        self.__write_raster__(output_file, dist, dx, [llc, urc],
+        self.__write_raster__(output_file, np.flipud(dist), dx, [llc, urc],
                               self.raster.ds.GetProjection())
