### 2018-12-19 ca6afe3
Making tests work with new real data
