+++ tests/test_interpolation.py
@@ -88,21 +88,18 @@ class TestRasterInterpolator(unittest.TestCase):
         raster_file = "tests/real_data/gebco_uk.tif"
         rci = RasterInterpolator(raster_file)
         rci.set_band()
-        points = ([823862., 5782011.],
-                  [839323., 5782408.],
-                  [853000., 5782804.],
-                  [858947., 5782606.],
-                  [866083., 5783201.],
-                  [889870., 5784787.],
-                  [949138., 5782408.],
+        points = ([842996., 5848009.],
+                  [834009., 5848207.],
+                  [832856., 5848273.],
+                  [828840., 5848306.],
+                  [823178., 5848503.],
                   )
-        expected = [-21.22,
-                    -24.03,
-                    -38.03,
-                    -46.72,
-                    -40.01,
-                    -32.13,
-                    -22.19]
+        expected = [-21.3,
+                    -25.3,
+                    -28.6,
+                    -13.5,
+                    -13.2,
+                    ]
         for p,e in zip(points, expected):
             self.assertAlmostEqual(rci.get_val(p),e,delta=0.75)
         
