### 2018-12-19 c856080
Update travis image
