### 2018-12-19 7f05e72
Fixing LINT errors
