### 2018-12-19 EnvModellingGroup/hrds#13 @jhill1
Fixed in #17
