### 2018-12-19 EnvModellingGroup/hrds#15 @jhill1
Fixed in #17
