### 2018-12-19 EnvModellingGroup/hrds#16 @jhill1
Fixed in #17
