### 2019-01-12 857b67f
fixed tests to find test data, even if run from outside the package dir.
Also PEP8 cleanup.
