+++ hrds/tests/test_hrds.py
@@ -17,9 +17,14 @@ import os
 #
 # Copyright Jon Hill, University of York, jon.hill@york.ac.uk
 
-base_raster = "tests/base.tif"
-layer1 = "tests/layer1.tif"
-layer2 = "tests/layer2.tif"
+# this is a bit kludgy, but should work on all platforms and test frameworks
+
+test_dir = os.path.split(__file__)[0]
+
+base_raster = os.path.join(test_dir, "base.tif")
+layer1 = os.path.join(test_dir, "layer1.tif")
+layer2 = os.path.join(test_dir, "layer2.tif")
+
 
 class TestHRDS(unittest.TestCase):
     """Tests the hrds.hrds.HRDS class"""
@@ -28,10 +33,10 @@ class TestHRDS(unittest.TestCase):
 
     def tearDown(self):
         # remove buffer files created
-        os.remove("tests/layer1_buffer.tif")
-        os.remove("tests/layer2_buffer.tif")
+        os.remove("layer1_buffer.tif")
+        os.remove("layer2_buffer.tif")
         return
-    
+
     def test_simple_rasters(self):
         """ Three layer test. Base layer is 100x100 with dx of (2.5,2).
             Origin is at (0,0)
@@ -44,18 +49,19 @@ class TestHRDS(unittest.TestCase):
         """
         bathy = HRDS(base_raster, rasters=(layer1, layer2), distances=(7, 5))
         bathy.set_bands()
-        points = ([5,5], # 1
-                  [40,50], # 3
-                  [28,32], # 2
-                  [13.5, 20], # 1.5
-                  [32.5, 45], # 2.5
+        points = ([5, 5],  # 1
+                  [40, 50],  # 3
+                  [28, 32],  # 2
+                  [13.5, 20],  # 1.5
+                  [32.5, 45],  # 2.5
                   )
         expected = [1.0, 3.0, 2.0, 1.5, 2.5]
-        for p,e in zip(points, expected):
-            self.assertAlmostEqual(bathy.get_val(p),e,delta=0.1)
+        for p, e in zip(points, expected):
+            self.assertAlmostEqual(bathy.get_val(p), e, delta=0.1)
 
-@unittest.skipUnless(os.path.isfile("tests/real_data/gebco_uk.tif"),
-                 "Skipping as proprietary data missing.")
+
+@unittest.skipUnless(os.path.isfile("real_data/gebco_uk.tif"),
+                     "Skipping as proprietary data missing.")
 class RealDataTest(unittest.TestCase):
 
     def test_real_world(self):
@@ -63,28 +69,28 @@ class RealDataTest(unittest.TestCase):
         Norfolk coast.
         Projected to UTM 30, so everything in m.
         """
-        bathy = HRDS("tests/real_data/gebco_uk.tif", 
-                     rasters=("tests/real_data/emod_utm.tif", 
-                              "tests/real_data/inspire_data.tif"), 
+        bathy = HRDS("real_data/gebco_uk.tif",
+                     rasters=("real_data/emod_utm.tif",
+                              "real_data/inspire_data.tif"),
                      distances=(700, 200))
         bathy.set_bands()
         """
         Our data:
         X	Y	emod_utm  gebco_uk  inspire_da
-        842996.	5848009.     	  -21.318	
-        834009.	5848207.	  -25.289	
-        832856.	5848273. -32.76	  -28.598	
-        828840.	5848306. -5.884	  -13.466	
-        823178.	5848503. -18.13	  -13.215	
+        842996.	5848009.     	  -21.318
+        834009.	5848207.	  -25.289
+        832856.	5848273. -32.76	  -28.598
[…]
