### 2019-01-12 EnvModellingGroup/hrds#20 @PythonCHB
Duplicate tests
There are two test classes called `RealDataTest` in `test_hrds.py`

I haven't checked if it's a complete duplicate by mistake, or an error -- but it should be fixed one way or another.
