### 2019-01-12 EnvModellingGroup/hrds#21 @PythonCHB
tests getting installed as a separate top level pacakge
The tests are put in a package that is next to the hrds package, and the setup.py uses setuptools.find_packages to to identify the packages to install. So when run, you get a top-level package called "tests" with the hrds tests in it. This is not good.

Here are some thoughts on where to put tests:

http://pythonchb.github.io/PythonTopics/where_to_put_tests.html

I've provided a PR here that puts the tests internal to the package:

#18

I also added formatting, etc changes to the setup.py in that PR.
