### 2019-01-12 EnvModellingGroup/hrds#21 @PythonCHB
NOTE: this could also be addressed by removing teh __init__.py, if you want to keep the tests separate from the package -- but with a tet suite this small, I think installing is is good.
