### 2019-01-14 EnvModellingGroup/hrds#23 @PythonCHB
Temp files creation.
In the `TestHRDS` class, the teardown removes two "buffer" files. But these are not explicitly created in the tests. I haven't dug into the code, but I assume these are automatically created by the HRDS class. Packages really shouldn't create non-explicit temp files like this -- either the user should explicitly specify where they want the temp files to be created, or they should be created in a temp dir (`import tempfile`), and ideally automatically deleted by the code.
