### 2019-01-14 76f3b41
re-restructure tests to be outside the pacakge.
__init__.py removed so that find_pacakges won't install it.
