### 2019-01-14 EnvModellingGroup/hrds#21 @PythonCHB
This is now fixed in #24
