### 2019-01-14 #1112 @PythonCHB
# Review of the HRDS software package:

https://github.com/EnvModellingGroup/hrds

Overall, this looks like a small, but useful package for raster data processing. But there are some issues with the pacakge structure, tests, and documentation, outlines below.

I recommend that the package be published once the below issues are addressed.

## Installation, etc:

### Platform support:

From the README:

"These instructions assume a Debian-based Linux. HDRS should work on other
systems, but is currently untested."

There are a lot of users out there that are not running Debian-based Linux -- this project should attempt to provide *some* guidance for other platforms. UNfotunately, the GDAL stack is pretty ugly to install, so perhaps recommending conda and conda-forge would be a good way to get people rolling in a platform independent way.

I notice that conda is (partially) being used for the CI -- so could full conda instuctions could be provided.

I am testing on OS-X with conda and conda-forge, Python 3.7

In fact, I highly recommend that the project authors make hrds available in conda-forge: https://conda-forge.org/#add_recipe

## Documentation and Examples:

In the main README, after providing Debian-only install instructions, the primary example is using the `firedrake` and `thetis` packages -- nifty use-case, but it would be much better, for a more general purpose package to have the primary examples be based only of the core package and its dependencies.

### Install

The hrds package is using a standard setuptools / setup.py install. However, there are some issues:

The tests are put in a package that is next to the hrds package, and the setup.py uses setuptools.find_packages to to identify the packages to install. So when run, you get a top-level package called "tests" with the hrds tests in it. This is not good.

Here are some thoughts on where to put tests:

http://pythonchb.github.io/PythonTopics/where_to_put_tests.html

I've provided a PR here that puts the tests internal to the package:

https://github.com/EnvModellingGroup/hrds/pull/18

I also added formatting, etc changes to the setup.py in that PR.

### tests:

* The way the test files were found was fragile -- so would only run if the current working dir was the parent of the tests dir:

https://github.com/EnvModellingGroup/hrds/issues/21

* There is a duplicate test name in: test_hrds.py -- those are skipped anyway due to lack of GEBCO data, but that should be fixed! Issue here:

https://github.com/EnvModellingGroup/hrds/issues/20

PR for fixes to the file finding, and PEP8 compliance fixes is here:

https://github.com/EnvModellingGroup/hrds/pull/24

### temp files.

The `HRDS` class appears to be creating temp buffer files that it does not delete:

https://github.com/EnvModellingGroup/hrds/issues/23

NOTE: computer memory is pretty huge these days -- are the buffer files even needed -- maybe only used when the rasters are really big?

## JOSS paper

The JOSS paper is fine as it goes, but there are a few issues:

* There are a few references cited, but nothing in the references section -- are those defined somewhere else?

* "a linear distance buffer is created to smoothly blend the two datasets together."

  I would like to see some description (and ideally a reference) for this algorithm.

* The docs do note that two rasters "are unlikely to
agree on the topographic/bathymetric height where they overlap", but there is no mention of (or algorithms for?) accounting for major differences -- different vertical datums, for instance.

## Documentation

The only user docs I could find is the README and docstrings. Some more comprehensive user and reference docs would be good. Also, the docstrings are not complete nor PEP 257 compliant.
