### 2019-01-15 #1112 @kthyng
Great! Thanks so much for the reviews, @edoddridge and @PythonCHB.

It looks like there are some unchecked check boxes for both reviews currently, and @jhill1 can work on addressing issues that have been brought up. Thanks again everyone.
