### 2019-01-15 #1112 @kthyng
Just a note that I will have an away message popping up starting tomorrow and I will continue to respond here while I can; once I can't and if this review is still on-going, then another editor will take over.
