### 2019-01-24 #1112 @jhill1
Thanks both for the reviews (especially whilst furloughed @PythonCHB - much appreciated).
