### 2019-01-27 #1112 @lheagy
:wave: Hi @jhill1, @edoddridge, @PythonCHB: just a quick note to say that I am happy to answer any questions or help out on the editor side while @kthyng is away.
