### 2019-04-09 2eb01f0
Remove buffer files by default and write temp files which get removed. Modified tests to check this works. Fixed duplicated test
