### 2019-04-09 EnvModellingGroup/hrds#20 @jhill1
Fixed in pull #26
