### 2019-04-09 EnvModellingGroup/hrds#23 @jhill1
Fixed in #26
