+++ docs/api.rst
@@ -1,0 +1,26 @@
+hrds API
+========
+
+hrds must be called within a python script. It has no user interface of its own.
+
+.. code-block:: python
+    
+    import hrds
+
+and all the functions below will be available hrds.xxxx.
+
+Below is a description of the functions that are available in the API.
+
+.. automodule:: hrds
+    :members:
+
+.. automodule:: hrds.hrds
+    :members:
+
+.. automodule:: hrds.buffer
+    :members:
+
+.. automodule:: hrds.RasterInterpolator
+    :members:
+
+
