### 2019-04-09 07e9065
Adding complete docstrings
