+++ hrds/raster.py
@@ -18,16 +18,26 @@ import math
 #
 # Copyright Jon Hill, University of York, jon.hill@york.ac.uk
 
+"""
+This module contains the Interpolator class and the RasterInterpolator
+class. These RasterInterpolator class reads in a GDAL raster and
+then allows integeragation of that data at an arbitrary point using
+bi-linear interoplation.
+"""
+
 
 class RasterInterpolatorError(Exception):
-    # any error generated by the RasterInterpolator object:
+    """
+    A generic error created by the RasterInterpolator class
+    """
 
     pass
 
 
 class CoordinateError(RasterInterpolatorError):
-    # error caused by coordinates being out of range or inside land mask:
-
+    """
+        Raised when a point is outside the raster or the land mask
+    """
     def __init__(self, message, x, i, j):
         self.message = message
         self.x = x
@@ -41,7 +51,7 @@ class CoordinateError(RasterInterpolatorError):
 class Interpolator(object):
     """Implements an object to interpolate values from a Raster-type data set.
 
-    Used by the RasterInterpolator below. A separate object as in future we
+    Used by the RasterInterpolator. A separate object as in future we
     may switch bands and hence have to reload the val data.
     """
 
@@ -53,9 +63,26 @@ class Interpolator(object):
         self.minmax = minmax
 
     def set_mask(self, mask):
+        """
+        Set a mask to use. Not yet implemented.
+        """
         self.mask = mask
 
     def get_val(self, point):
+        """
+        Get the value of this raster at the desired point via bi-linear 
+        interpolation.
+
+        Args:
+            point: a length 2 list containing x,y coordinates
+
+        Returns:
+            The value of the raster stack at that point
+
+        Raises:
+            CoordinateError: The point is outside the rasters
+            RasterInterpolatorError: Generic error interpolating data at that point
+        """        
         yhat = ((point[0]+(self.delta[0]/2.0)-self.origin[0])/self.delta[0])
         xhat = ((point[1]+(self.delta[1]/2.0)-self.origin[1])/self.delta[1])
         j = int(math.floor(yhat))-1
@@ -119,7 +146,8 @@ class Interpolator(object):
 # the latter is considered immutable, whereas the NetCDFInterpolator may
 # change in future
 class RasterInterpolator(object):
-    """Implements an object to interpolate values from a Raster-stored data set.
+    """
+    Implements an object to interpolate values from a Raster-stored data set::
 
         rci = RasterInterpolator('foo.tif')
 
@@ -127,13 +155,13 @@ class RasterInterpolator(object):
     to be the lower-left corner (i.e. south west)
     and the projection space is stored with the raster.
 
-    To indicate the band to be interpolated:
+    To indicate the band to be interpolated::
 
         rci.set_band(2)
 
-    The deafut is Band 1 (ie. with no number given)
+    The default is Band 1 (ie. with no number given)
 
-    To interpolate this field in any arbitrary point:
+    To interpolate this field in any arbitrary point::
 
         rci.get_val((-3.0, 58.5))
 
@@ -157,14 +185,9 @@ class RasterInterpolator(object):
     def get_extent(self):
         """Return list of corner coordinates from a geotransform
 
-            @type gt:   C{tuple/list}
-            @param gt: geotransform
-            @type cols:   C{int}
-            @param cols: number of columns in the dataset
-            @type rows:   C{int}
-            @param rows: number of rows in the dataset
-            @rtype:    C{[float,...,float]}
-            @return:   coordinates of each corner
+        Returns:
+            List continaing the corner coordinates of the raster
+
         """
         cols = self.ds.RasterXSize
         rows = self.ds.RasterYSize
[…]
