+++ hrds/__init__.py
@@ -14,13 +14,11 @@
 # Copyright Jon Hill, University of York, jon.hill@york.ac.uk
 
 """
-Python package for Hierarchical raster data sets. Main components are:
+hrds contains the following components:
 
-    buffer - functions to create buffer files
-
-    hdrs - the main hdrs object.
-
-    RasterInterpolator - objects to interpolate individual rasters
+    * buffer - functions to create buffer files
+    * hdrs - the main hdrs object.
+    * RasterInterpolator - objects to interpolate individual rasters
 
 """
 
