+++ setup.py
@@ -28,7 +28,7 @@ setuptools.setup(name='hrds',
                  author='Jon Hill',
                  author_email='jon.hill@york.ac.uk',
                  description=DESCRIPTION,
-                 url='https://github.com/stephankramer/uptide',
+                 url='https://github.com/EnvModellingGroup/hrds',
                  packages=setuptools.find_packages(),
                  keywords=['raster', 'gis', 'modelling'],
                  classifiers=[
