### 2019-04-09 d8b86ed
Fixing lint errors and adding correct formatting to sphinx docs
