### 2019-04-09 f1a4f8d
Minor tweaks to docs and adding user doc to docs folder
