### 2019-04-10 #1112 @kthyng
Great @jhill1!

@PythonCHB can you look through @jhill1's work to see if you are satisfied to check off the remaining boxes in your review, or if you have follow up comments?
