### 2019-05-04 4878b99
small typo in setup.py
