### 2019-05-04 #1112 @PythonCHB
Thanks for addressing my concerns -- this is a more robust package now!

A few issues still:

## The Readme:

I have copy edited the README -- PR here:

https://github.com/EnvModellingGroup/hrds/pull/29

Most of that is just copy editing, but a small issue -- in the example in the readme, there is this code:

```
import sys
sys.path.insert(0,"../../")

from hrds import hrds
```

messing around with `sys.path` like that is not considered good practice these days. As the previous instructions are how to install the package, this is totally unnecessary. And if a user wants to work with the package directly from the source, the recommended way to do that is an "editable install":

`pip install -e ./`

and then the package will be available with a regular import.

I have removed that code in my PR.

In the `thetis` example, there is:

```
from firedrake import *
from thetis import *
from firedrake import Expression
```
`import *` is highly discouraged these days -- and particularly in example code like this, it's then confusing be3cause you don't know if a name comes from firedrake or thetis. And if you have imported everything from firedrake, you don't need to import Expression.

`import *` should be removed -- I have not done this in my PR, as I haven't installed thetis and firedrake to test.

## Documentation

I've added a few copy-edits tot he PR above.

Note that there is a fair bit of overlap between the README and the Sphinx Docs -- (and teh Sphinx docs now need copy editing to match the README). I suggest that you single source these if you can.

gitHub will allow READMEs in RST as well, so you could have one RST page with the intro and all,  and include that into your Sphinx docs.

You could also make the README smaller, and link to the more substantial docs -- which you could publish as html via gitHubIO or readthedocs.

### docstrings and the API docs:

It looks like Sphinx auto-doc is not including teh class __init__ docs -- which can be very helpful.

Also, some of those __init__ docstrings (hrds.hrds.HRDS.__init__, for instance, are not complete -- you should document all the parameters to the __init__.

## minor notes:

`from __future__ import print_function` is necessary for py3 only code.

No need to derive from `object` in py3 either.

### Code style:

I see this:

```
    def set_mask(self, mask):
        """
        Set a mask to use. Not yet implemented.
        """
        self.mask = mask
```

This is why Python has "properties" -- there is not need to write do-nothing getters and setters.

https://uwpce-pythoncert.github.io/PythonCertDevel/modules/Properties.html?highlight=properties

## The Paper

The References section is blank -- but there is at least one reference in the paper -- am I missing something?

I'd still like to see a more description the buffering algorithm -- linear interpolation does not need defining, but exactly how one interpolates between two different resolutions of rasters in the buffer zone is not obvious.

# Conclusion:

None of these issues are fatal -- congrats! You've made a nice contribution to the community.
