### 2019-05-08 #1112 @kthyng
Thanks so much @PythonCHB!

@jhill1 Can you address these comments soon?
