### 2019-05-09 c096f30
Fixing example thetis code to remove import *
