+++ hrds/raster_buffer.py
@@ -1,4 +1,3 @@
-from __future__ import print_function
 import numpy as np
 from osgeo import gdal
 from .raster import RasterInterpolator
