### 2019-05-09 b1b86aa
Adding figure to JOSS paper
