### 2019-05-09 fa204cc
Adding data points to figure
