### 2019-05-09 #1112 @jhill1
Thanks @PythonCHB !

The pull request has been merged (thanks for that!).

**Documents**: I've pulled over the changes in the README to the Sphinx docs and will have a think about how to make the best of them. I have a summer student this year who will be using this code, so we'll see how they get on and then I might ask them to rationalise the docs along the lines proposed.

I've removed all the `import *` from docs and code examples and the leftover Python 2 code. I haven't removed the set_mask yet (as a reminder for me to implement this) - the code will have to do more than set the self.mask variable (e.g. regenerate buffers).

**Paper:** I've added a figure to explain the buffer which will hopefully help The references are added automatically by JOSS when the PDF is created.
