### 2019-05-09 49d4069
Fixing minor issues with figure
