### 2019-05-09 #1112 @kthyng
Great work @jhill1 it looks like you have addressed all of @PythonCHB's comments. There is one unchecked box, but the pdf does indeed now have references included and they have doi's when appropriate.

I am going through my own checks now.
