### 2019-05-09 #1112 @kthyng
@jhill1 your citation of Martin-Short2015 shouldn't have parentheses so that it is inline as it is part of the sentence as written.
