### 2019-05-09 cb87ea5
Minor fixes and update version
