### 2019-05-09 #1112 @kthyng
@jhill1 agreed!

Ok I just looked up [your submission to Zenodo](https://zenodo.org/record/2691685#.XNSRrpNKgWo) and the metadata includes @PythonCHB which is nice but not correct — this often happens. Can you edit the metadata so that it matches your paper submission? In particular you want the author list to match, and the title to match.
