### 2019-05-10 #1112 @PythonCHB
Not really a big deal: but if you are publishing this, maybe it's time for version 1.0?

If it's not really ready for 1.0, it's not ready for publication, either.
