### 2019-05-13 #1112 @jhill1
Leave for now, please. @PythonCHB is right that it should have been made v1.0, but I've already pushed pip and conda updates with 0.12 and referred to the version being the one published. I'd rather keep a cleaner history.
