### 2019-05-13 #1112 @labarba
@PythonCHB — you have the "References" item unchecked. Did you find something missing there?
