### 2019-05-13 #1112 @labarba
Congratulations, @jhill1, your JOSS paper is published 🚀

Sincere thanks to our editor: @kthyng, and the reviewers: @edoddridge, @PythonCHB — we appreciate your enormous contribution! 🙏
